refactoring application models structure
