+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/shared_handlers.py
@@ -3,8 +3,8 @@ from typing import Optional, Tuple
 
 from fastapi import FastAPI
 
+from .models.domains.shared_store import SharedStore
 from .settings import DynamicSidecarSettings
-from .shared_store import SharedStore
 from .utils import async_command, write_to_tmp_file
 
 logger = logging.getLogger(__name__)
