+++ services/dynamic-sidecar/tests/conftest.py
@@ -13,9 +13,9 @@ from async_asgi_testclient import TestClient
 from fastapi import FastAPI
 from pytest_mock.plugin import MockerFixture
 from simcore_service_dynamic_sidecar.application import assemble_application
+from simcore_service_dynamic_sidecar.models.domains.shared_store import SharedStore
 from simcore_service_dynamic_sidecar.settings import DynamicSidecarSettings
 from simcore_service_dynamic_sidecar.shared_handlers import write_file_and_run_command
-from simcore_service_dynamic_sidecar.shared_store import SharedStore
 
 
 @pytest.fixture(scope="module", autouse=True)
